# Intro to Hand Tools In the next section of the Tools documentation for construction, we will be examining various hand tools. The purpose of this section is for the reader to understand safe steps on how to use these tools efficiently for robotics work. It is important to read carefully about the tips and tricks for these tools in order to minimize the chance of injury when working on the robot!
### Must-Read for All Tools
It is important that you don't use a tool to do tasks that it's not meant to be used for. For example, using a screwdriver as a mallet. Also, it's important to always remember to return tools back to the desinated areas after use before leaving the Maker Space.

### Screwdrivers
There are two types of screwdrivers: phillips head and flathead. Always make sure to use the proper screwdriver on a screw. Use the screwdriver to take out and put in screws on a mechanism. An example is taking off bumper screws on a robot.

### Allen Key
An allen key is used to drive bolts and screws with hexagonal sockets in their heads.

### Center Punch
Center punches come in different sizes. They are used to mark the initial dent/hole before drilling it. After defining the horizontal and vertical, hold the pointed end of a center punch on the cross-section and hit the flat end with a mallet. This will create a mark, which you will drill out and/or drive a fastener into.

### Scribe
DO NOT use the scribe as a center punch. Scribes are much longer and have sharper points, which are used to mark lines/outlines of a metal piece that are going to be cut. A sharpie works as well, but a scribe is better.

### Must-Read for Saws
For both saws, make sure to use the entire blade to cut and don’t saw in only one place because this would cause inconsistencies on the blade for the teeth would wear down faster than the other areas.

### Hand Saw
The wood hand saw is used to cut wood, plywood, 2x4, and etc.

### Hacksaw
The metal hand saw is used to cut small metal pieces, aluminum churros, ThunderHex, and etc.

### Pliers
Pliers are used to hold small objects. It’s function might vary depending on the type of plier. An example is removing hot bits. A type of plier is the gearbox snap ring.

### Wrench
The wrench is use to grip, fasten, turn, tighten, and loosen things like locknuts, nuts, and bolts. It comes in different sizes.

### Mallets
Mallets are mainly used to hit the center punch. Mallets are also used to apply force to parts to shape or move them with less risk of damaging parts. For example, applying force with the mallet to a bearing to get it into it’s frame, then using the mallet to insert a hexshaft through the bearing(s).

## Rivets Guns
Rivet guns are used to drive rivets in to fasten sheet metal or heavy fabric. The thicker side of the rivet goes into the hole, and the slimmer side goes into the rivet gun. When using the rivet gun, keep the rivet and rivet gun straight against the surface. To remove a rivet, drill it out.