# Intro to Fasteners
Welcome to the tools section of the construction documentation! For the first part of this section we will be taking a look at fasteners. Fasteners are essential for locking in paramount pieces that allow the robot to maintain form. Each type of fastener comes in many forms, shapes and sizes; while we may not cover every one of them in detail, we hope that this documentation gives you a strong foundation with using construction tools.

[For a quick look at a variety of fasteners!](https://www.boltdepot.com/fastener-information/Printable-Tools/Type-Chart.pdf)

### Screws
Little stripping of screws is acceptable. Remember to scrap any type of stripped screws, nuts, or bolts.

### Bolts
Bolts are not used often here in robotics. You can read more about them here.

### Locknuts
Read about them in teh linked doc. There's not much to say about this.

### Nuts
Only use the wrench to grip nuts and bolts.

### Washer
Washers should be used with screws, bolts, and nuts to prevent loosing and to distribute the weight between the washer and screws, bolts, or nuts over a larger area.

### Spacers
Spacers come in different sizes (1/8", 1/4", 2", etc.). Acetal Hex Spacers are used to maintain space between components.

### Locite(glue)
Locite is used to lock a screw in place. It has three different strengths. Blue is the lowest strength. Red is the medium strength. Green is the strongest strength.